Added Interface ConstraintSpace implemented by ReifiedSpace Added modfdb.hh Modified modsearch.hh
